# spec/openapi/uai-v1.yaml
openapi: 3.1.0
info:
  title: Universal Agent Identity API
  version: "0.1.0"
  summary: Identity, attestation and governance for autonomous AI agents.
  description: |
    The UAI v1 REST surface.

    REST and JSON because a protocol meant to be implemented by everyone must be
    implementable with `curl`. gRPC is used internally where call volume justifies it.

    **Authentication.** Every state-changing call requires proof of possession: mutual TLS
    with a SPIFFE X.509-SVID *and* an RFC 9421 HTTP message signature. mTLS proves the
    runtime instance; the message signature proves the persistent identity. Either alone is
    insufficient, because a stolen SVID has no persistent key and a stolen persistent key has
    no attested runtime.

    **What this API does not claim.** A verified identity is not a safe agent, and revocation
    does not stop code from running outside the ecosystem. See
    `docs/protocol/01-overview.md`.
  license:
    name: Apache-2.0 (implementation) / CC-BY-4.0 (specification)
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: UAI security
    email: security@uai.world
servers:
  - url: https://api.uai.world/v1
    description: Production API
  - url: https://verify.uai.world/v1
    description: Public verification endpoint (unauthenticated, cached)
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - mutualTLS: []
    messageSignature: []
paths:
  /agents:
    post:
      tags: [identity]
      operationId: registerAgent
      summary: Begin agent registration
      description: |
        Returns challenges for BOTH the owner and the agent. Ownership is proven, not declared:
        the two signatures must reference the same (agent, owner) pair, so nobody can claim
        another party's agent and no owner can disown an agent retroactively.

        UAI never generates an agent's private key. A registry that can generate your key can
        impersonate you, which would void the entire attribution claim.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegistrationRequest' }
      responses:
        '202':
          description: Registration opened; proofs required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistrationChallenge' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /agents/{id}/prove:
    post:
      tags: [identity]
      operationId: submitRegistrationProof
      summary: Submit an owner or agent proof of possession
      description: |
        The path identifier is a REGISTRATION id, not a UAI-ID. No UAI-ID exists at this point in
        the flow: the ULID is minted only after both proofs verify (§8.2), so an agent identifier
        here would name something that does not yet exist. The `reg_` prefix makes the two
        impossible to confuse.

        Each half resolves its verification key differently, and the asymmetry is the security
        argument. The agent's key arrives in `public_jwk` because it is being introduced. The
        owner's key is read from the registry and MUST NOT be sent: a key supplied by the caller
        would make ownership self-asserted, which is exactly what §6.4.1 exists to prevent.
      parameters:
        - $ref: '#/components/parameters/RegistrationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProofSubmission' }
      responses:
        '201':
          description: Identity minted once both proofs are present
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisteredAgent' }
        '202':
          description: Proof accepted; the other half is still outstanding
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PendingRegistration' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /agents/{id}:
    get:
      tags: [identity]
      operationId: getAgent
      summary: Universal Agent Identity Card
      parameters: [{ $ref: '#/components/parameters/AgentId' }]
      responses:
        '200':
          description: Identity card
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisteredAgent' }
        '404': { $ref: '#/components/responses/Problem' }
  /agents/{id}/credentials:
    get:
      tags: [identity]
      operationId: getAgentCredentials
      summary: Credential set held by an agent
      parameters: [{ $ref: '#/components/parameters/AgentId' }]
      description: |
        Credentials are returned exactly as issued, byte for byte. Re-serializing them would
        produce different bytes than the issuer signed and every proof would fail, for reasons
        that look like a key problem.

        An AgentOwnershipCredential embeds the two-sided registration proof, so a relying party
        validates ownership from the document and the owner's public key alone (§6.4.1).
      responses:
        '200':
          description: Credentials
          content:
            application/json:
              schema:
                type: object
                required: [credentials]
                properties:
                  credentials:
                    type: array
                    description: |
                      W3C Verifiable Credentials 2.0 with Data Integrity proofs. See
                      spec/schemas/agent-identity-credential.schema.json and
                      agent-ownership-credential.schema.json.
                    items: { type: object }
        '404': { $ref: '#/components/responses/Problem' }
  /agents/{id}/bind:
    post:
      tags: [identity]
      operationId: bindAgent
      summary: Bind a runtime instance to a persistent identity
      description: |
        The signature covers the challenge AND the SVID identifiers, which is what
        cryptographically ties "who I am" to "where I am running". Either alone would be
        forgeable by an attacker holding the other.
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BindRequest' }
      responses:
        '202':
          description: |
            Challenge issued. Sent with an empty body, these endpoints ask for a challenge; the
            agent cannot sign a binding statement until the registry has told it which value to
            sign over (§9.1).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BindingChallenge' }
        '200':
          description: Bound
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BindResult' }
        '409': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /agents/{id}/unbind:
    post:
      tags: [identity]
      operationId: unbindAgent
      summary: Leave UAI (self-service)
      description: |
        An agent may leave at any time without asking anyone. This is a security property as
        much as a principle: a protocol you cannot leave is one operators refuse to adopt.
        Unbinding is not deletion — every existing attestation stays verifiable forever.
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnbindRequest' }
      responses:
        '202':
          description: |
            Challenge issued. Sent with an empty body, these endpoints ask for a challenge; the
            agent cannot sign a binding statement until the registry has told it which value to
            sign over (§9.1).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BindingChallenge' }
        '200':
          description: Unbound; history retained
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BindResult' }
        '409': { $ref: '#/components/responses/Problem' }
  /agents/{id}/rebind:
    post:
      tags: [identity]
      operationId: rebindAgent
      summary: Rejoin UAI with cryptographic continuity
      description: |
        The continuity proof MUST be produced by a key that was valid at the moment of
        unbinding. Without that rule, "unbind, rotate key, rebind" would be a laundering path
        for a stolen identity.
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RebindRequest' }
      responses:
        '202':
          description: |
            Challenge issued. Sent with an empty body, these endpoints ask for a challenge; the
            agent cannot sign a binding statement until the registry has told it which value to
            sign over (§9.1).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BindingChallenge' }
        '200':
          description: Rebound
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BindResult' }
        '409': { $ref: '#/components/responses/Problem' }
        '400': { $ref: '#/components/responses/Problem' }
  /agents/{id}/events:
    get:
      tags: [action]
      operationId: getAgentEvents
      summary: The agent's event chain, with proofs
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        '200':
          description: Chain page
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/ActionAttestation' }
                  next_cursor: { type: string }
  /passports/request:
    post:
      tags: [passport]
      operationId: requestPassport
      summary: Request a cross-jurisdictional passport
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PassportRequest' }
      responses:
        '201':
          description: Passport issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentPassportCredential' }
        '202':
          description: Held for human approval
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /passports/{id}:
    get:
      tags: [passport]
      operationId: getPassport
      summary: Fetch a passport credential
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, pattern: '^urn:uai:passport:[0-9A-HJKMNP-TV-Z]{26}$' }
      responses:
        '200':
          description: Passport
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentPassportCredential' }
        '404': { $ref: '#/components/responses/Problem' }
  /passports/check:
    get:
      tags: [passport]
      operationId: checkPassport
      summary: Run the passport checklist for a capability and a jurisdiction set
      description: |
        The 11.6 checklist, answered for anyone who asks. A GET because it changes nothing,
        and unauthenticated because the party who needs this answer is the one being dealt
        with: an answer only the agent itself could obtain would be useless to them.

        A passport in scope does NOT authorize an action. The guardrail still evaluates it.
        `code` is `UAI_PASSPORT_REQUIRED` when the agent has no passport at all, which is a
        different fact from a passport that refuses and must not be rendered as one.
      parameters:
        - name: subject
          in: query
          required: true
          description: The UAI-ID being checked
          schema: { type: string }
        - name: capability
          in: query
          required: true
          schema: { type: string }
        - name: targets
          in: query
          required: true
          description: Comma-separated ISO 3166-1 alpha-2 codes
          schema: { type: string }
        - name: risk_class
          in: query
          schema: { enum: [INFORMATIONAL, LOW, MODERATE, HIGH, CRITICAL] }
      responses:
        '200':
          description: Check result
          content:
            application/json:
              schema:
                type: object
                required: [subject, capability, effect, allowed, reason, note]
                properties:
                  subject: { type: string }
                  capability: { type: string }
                  targets: { type: array, items: { type: string } }
                  effect:
                    enum: [ALLOW, ALLOW_WITH_MONITORING, REQUIRE_HUMAN_APPROVAL, DENY]
                  allowed: { type: boolean }
                  code: { type: string, description: A UAI_* code when the effect is DENY }
                  reason: { type: string }
                  passport_id: { type: string }
                  state_at_decision: { type: string }
                  credential_hash: { type: string }
                  policy_version: { type: string }
                  note: { type: string }
  /capability-requests:
    post:
      tags: [action]
      operationId: requestCapability
      summary: Request a capability the agent does not hold
      description: |
        Creates a PENDING request and nothing else. It does not grant, and it cannot be made
        to: approval is an act by the human owner, out of band. There is no route in this API
        that writes a capability grant, and the database refuses one signed by the agent
        itself (22.9, T-11/T-13).

        Requires an RFC 9421 signature tagged `UAI-v1:capability-request`. One open request per
        (agent, capability): re-asking in a loop until an owner approves out of fatigue is a
        social attack on a human, and it is refused with 409.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capability, justification]
              properties:
                capability: { type: string }
                justification:
                  type: string
                  description: |
                    What it is for, in terms the owner can decide on. Required: without it they
                    would be approving a capability name.
                purpose: { type: string }
                resource: { type: string }
      responses:
        '202':
          description: Request recorded, pending the owner's decision
          content:
            application/json:
              schema:
                type: object
                required: [request_id, capability, state, granted, note]
                properties:
                  request_id: { type: string }
                  capability: { type: string }
                  state: { enum: [PENDING] }
                  granted:
                    type: boolean
                    enum: [false]
                    description: Always false. Present so a caller reading this sees the answer.
                  requested_at: { type: string, format: date-time }
                  expires_at: { type: string, format: date-time }
                  note: { type: string }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /agents/{id}/capability-requests:
    get:
      tags: [action]
      operationId: listCapabilityRequests
      summary: An agent's capability requests
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent: { type: string }
                  requests: { type: array, items: { type: object } }
  /policy/evaluate:
    post:
      tags: [action]
      operationId: evaluatePolicy
      summary: Policy decision, BEFORE the action is executed
      description: |
        A decision taken after execution would be theater. Every decision is recorded,
        including ALLOW: a guardrail that logs only denials cannot answer what was permitted
        and why, which is the question that matters after an incident.

        Fails closed for identity, capability and passport checks. If the policy service is
        degraded it uses the last signed cached bundle and says so in the record.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PolicyEvaluationRequest' }
      responses:
        '200':
          description: Decision record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicyDecision' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
  /actions/{eventId}:
    get:
      tags: [action]
      operationId: getAction
      summary: An attestation with its transparency receipt
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$' }
      responses:
        '200':
          description: Attestation and proof
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttestationReceipt' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: SPIFFE X.509-SVID. Proves the runtime instance.
    messageSignature:
      type: apiKey
      in: header
      name: Signature
      description: |
        RFC 9421 HTTP message signature over @method, @target-uri, content-digest,
        uai-agent-id and uai-nonce, tagged with the UAI domain. Proves the persistent
        identity. `agent_id` in a JSON body proves nothing.
    webAuthn:
      type: apiKey
      in: header
      name: X-UAI-Assertion
      description: |
        WebAuthn assertion from a hardware authenticator with user verification required.
        For votes, the challenge is the digest of the vote statement itself.
...